+++ compressors/gfxcomp_zx0.cpp
@@ -20,13 +20,28 @@ extern "C" __declspec(dllexport) const char* getExt()
 }
 
 // The actual compressor function, calling into the zx0 code
-uint32_t compress(const uint8_t* pSource, const size_t sourceLength, uint8_t* pDestination, const size_t destinationLength)
+uint32_t compress(
+    const uint8_t* pSource,
+    const size_t sourceLength,
+    uint8_t* pDestination,
+    const size_t destinationLength)
 {
     int outputSize;
     int delta; // we don't care about this
-    const auto optimised = optimize(const_cast<unsigned char*>(pSource), sourceLength, 0, sourceLength);
+    const auto optimised = optimize(
+        const_cast<unsigned char*>(pSource),
+        static_cast<int>(sourceLength),
+        0,
+        static_cast<int>(sourceLength));
     // ZX0 leaks data by design, we can't free this
-    const auto pOutputData = compress(optimised, const_cast<unsigned char*>(pSource), sourceLength, 0, 0, &outputSize, &delta);
+    const auto pOutputData = compress(
+        optimised,
+        const_cast<unsigned char*>(pSource),
+        static_cast<int>(sourceLength),
+        0,
+        0,
+        &outputSize,
+        &delta);
 
     if (static_cast<uint32_t>(outputSize) > destinationLength)
     {
@@ -41,12 +56,21 @@ uint32_t compress(const uint8_t* pSource, const size_t sourceLength, uint8_t* pD
     return outputSize;
 }
 
-extern "C" __declspec(dllexport) int compressTiles(const uint8_t* pSource, const uint32_t numTiles, uint8_t* pDestination, const uint32_t destinationLength)
+extern "C" __declspec(dllexport) int compressTiles(
+    const uint8_t* pSource,
+    const uint32_t numTiles,
+    uint8_t* pDestination,
+    const uint32_t destinationLength)
 {
-    return compress(pSource, numTiles * 32, pDestination, destinationLength);
+    return static_cast<int>(compress(pSource, numTiles * 32, pDestination, destinationLength));
 }
 
-extern "C" __declspec(dllexport) int compressTilemap(const uint8_t* pSource, const uint32_t width, const uint32_t height, uint8_t* pDestination, const uint32_t destinationLength)
+extern "C" __declspec(dllexport) int compressTilemap(
+    const uint8_t* pSource,
+    const uint32_t width,
+    const uint32_t height,
+    uint8_t* pDestination,
+    const uint32_t destinationLength)
 {
-    return compress(pSource, width * height * 2, pDestination, destinationLength);
+    return static_cast<int>(compress(pSource, width * height * 2, pDestination, destinationLength));
 }
